refactor bridge blc test
