+++ contracts/bridge/e2e/helper/helper.js
@@ -121,5 +121,5 @@ module.exports = {
   createCompleteValidatorProofs,
   createRandomPositions,
   printBitfield,
-  roundUpToPow2
+  roundUpToPow2,
 }
